Reverse Linked List Proram
This program reverses a singly linked list using three points. This is the only time that I've had to look up a coding solution online so this program haunts me to this day. One day I'll have to go back and reattempt this program from scratch to prove to myself that I can do it. I will never forget the hours of frustration this program left me with.