Struct Vector Program
Implements two structures one of which is used as a vector within the other. Then it fills in data from an external file and then implements a function to see if a student is in the external file and also a function that gets the letter grade for a course.