Specialized Template Programs
This is probably one of my favorite programs that I have every made because I learned so much from it. Basically I implemented a generic functions that finds the highest of two identical data types yet it wouldn't work with some strings for some reasons. I realized that this is because when you pass a string to a function such as "string example" is actually passes a constant character array so its case has to be considered a little more specially. To fix this you can make a specialized template that is for this specific data type that allows you to evaluated the specific type correctly also overloaded operators can be used in some cases as well.